pyproject.toml
[tool.poetry]
name = "python-startup"
version = "0.1.0"
description = ""
authors = ["Adnan <adnankattan9@gmail.com>"]
requires-python = "^3.11"

[tool.poetry.dependencies]
qdrant-client = "~=1.11.1"
fasttext-langdetect = { version = "~=1.0.5", markers = 'sys_platform != "win32"' }
SQLAlchemy = "~=2.0.31"
alembic = "~=1.13.2"
fastapi = "~=0.114.2"
pydantic = "~=2.9.0"
starlette = "~=0.38.2"
uvicorn = "~=0.30.6"
uvicorn-worker = "~=0.2.0"
httptools = "~=0.6.1"
pyyaml = "~=6.0.1"
uvloop = { version = "~=0.20.0", markers = 'sys_platform != "win32"' }
email-validator = "~=2.1.1"
stringcase = "~=1.2.0"
python-jose = "~=3.3.0"
passlib = "~=1.7.4"
pytest = "~=8.2.0"
psutil = "~=5.9.8"
requests = "~=2.32.3"
phonenumbers = "~=8.13.36"
redis = "~=5.0.4"
pika = "~=1.3.2"
python-socketio = "~=5.11.2"
websockets = "~=12.0"
websocket-client = "~=1.8.0"
aiohttp = "~=3.9.5"
aiodns = "~=3.1.1"
aio-pika = "~=9.4.1"
beanie = "~=1.26.0"
motor = "~=3.4.0"
pymongo = "~=4.7.2"
pydantic-settings = "~=2.4.0"
# async-fastapi-jwt-auth = "~=0.6.5" # Due to conflicting versions
PyJWT = "~=2.8.0"
httpx = "~=0.27.0"
cryptography = "~=42.0.2"
gunicorn = "~=22.0.0"
celery = "~=5.4.0"
python-multipart = "~=0.0.9"
sentry-sdk = { version = "~=2.1.1", extras = ["fastapi"] }
typing_extensions = "~=4.11.0"
pydantic_core = "~=2.23.2"
srt = "~=3.5.3"
toolz = "~=0.12.1"
aiostream = "~=0.6.2"
asyncache = "~=0.3.1"
cachetools = "~=5.4.0"
orjson = "~=3.10.3"
redisvl = "~=0.1.3"
dishka = "~=1.3.0"
watchfiles = "~=0.23.0"
faststream = "~=0.5.19"
fsspec = "~=2024.9.0"
fastapi-mail = "~=1.4.1"
firebase_admin = "~=6.5.0"
s3fs = "~=2024.9.0"
gcsfs = "~=2024.9.0.post1"
pillow = "~=10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
